Re-queue DigitalObjectImport if given parent is not found
The parent of a Digital Object can be specified at the time of import.
However, if the specified parent is not found, the Digital Object
cannot be imported. However, it is possible that the parent Digital
Object is in the queue, ready to be imported, but has not yet been
processed. Therefore, the child Digital Object will be re-queued,
thus allowing the parent Digital Object to be processed before the
re-queued child Digital Object.
Digital Objects can be re-queued at most MAX_REQUEUES=3 times
Note that DigitalObjectImport instances are used to represent the
Digital Objects waiting to be imported.