Add #update_doi_metadata to DigitalObject::Ezid
This method will be when a DigitalObject already has a DOI EZID (possibly
from a previous publish) and is subsequently (re)published. The
metadata for the DigitalObject may have changed, and the pertinent
metadata on the EZID server should be updated. This method should
be called to perform the update.